refactor: siteBuilder.js file Created
And set-up on other files to accommodate upcoming refactor.
siteBuilder.jst will use model.js data to populate index.html added
with structure and content.

The end result will be seperation of concerns, and an MVVM designed web
site, that reduces redundancy and scales well for further Apps of be
developed.